+++ src/components/common/ErrorPage.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useRouter } from 'next/navigation';
-import MoonoError from '@/assets/icon/moono_lost.svg?react';
+import MoonoError from '@/assets/moono/moono_lost.svg?react';
 import Button from '@/components/common/Button';
 
 interface ErrorPageProps {
@@ -10,37 +10,40 @@ interface ErrorPageProps {
   reset: () => void;
 }
 
-export default function ErrorPage({ message, buttonColor, reset }: ErrorPageProps) {
-
+export default function ErrorPage({
+  message,
+  buttonColor,
+  reset,
+}: ErrorPageProps) {
   const router = useRouter();
-  
+
   const ClickReset = () => {
     reset();
     router.replace('/');
   };
-  
+
   return (
-    <div className="flex flex-col items-center min-h-screen bg-white pt-[205px]">
-    
+    <div className="flex min-h-screen flex-col items-center bg-white pt-[205px]">
       {/* 무너 우는 이미지 */}
       <div className="mb-[35px] flex justify-center">
-        <MoonoError aria-hidden="true"
-        />
+        <MoonoError aria-hidden="true" />
       </div>
 
       {/* 에러 안내 텍스트 */}
-      <div className="text-center px-[74px]">
-        <p className="script-title whitespace-pre-wrap">
-            {message}
-        </p>
+      <div className="px-[74px] text-center">
+        <p className="script-title whitespace-pre-wrap">{message}</p>
       </div>
 
       {/* 버튼 */}
-      <div className="mt-[80px] px-[45px] w-full flex justify-center">
+      <div className="mt-[80px] flex w-full justify-center px-[45px]">
         <Button
           size="full"
           variant="solid"
-          className={buttonColor ? "text-[`#111111`] script-title w-full" : "bg-primary script-title w-full"}
+          className={
+            buttonColor
+              ? 'script-title w-full text-[`#111111`]'
+              : 'bg-primary script-title w-full'
+          }
           style={buttonColor ? { backgroundColor: buttonColor } : {}}
           onClick={ClickReset}
         >
@@ -48,5 +51,5 @@ export default function ErrorPage({ message, buttonColor, reset }: ErrorPageProp
         </Button>
       </div>
     </div>
-  )
+  );
 }
